split validatefuncs into own file and fix lint errors
