+++ vcd/resource_vcd_lb_service_monitor.go
@@ -102,33 +102,6 @@ func resourceVcdLbServiceMonitor() *schema.Resource {
 	}
 }
 
-// validateCase checks if a string is of caseType "upper" or "lower"
-func validateCase(caseType string) schema.SchemaValidateFunc {
-	return func(i interface{}, k string) (s []string, es []error) {
-		v, ok := i.(string)
-		if !ok {
-			es = append(es, fmt.Errorf("expected type of %s to be string", k))
-			return
-		}
-
-		switch caseType {
-		case "upper":
-			if strings.ToUpper(v) != v {
-				es = append(es, fmt.Errorf(
-					"expected string to be upper cased, got: %s", v))
-			}
-		case "lower":
-			if strings.ToLower(v) != v {
-				es = append(es, fmt.Errorf(
-					"expected string to be lower cased, got: %s", v))
-			}
-		default:
-			panic("unsupported validation type for validateCase() function")
-		}
-		return
-	}
-}
-
 func resourceVcdLbServiceMonitorCreate(d *schema.ResourceData, meta interface{}) error {
 	vcdClient := meta.(*VCDClient)
 	vcdClient.lockParentEdgeGtw(d)
