+++ vcd/suppressfuncs.go
@@ -2,7 +2,6 @@ package vcd
 
 import (
 	"flag"
-	"fmt"
 	"net"
 	"os"
 
@@ -17,42 +16,6 @@ func suppressWordToEmptyString(word string) schema.SchemaDiffSuppressFunc {
 			return true
 		}
 		return false
-	}
-}
-
-// noopValueWarningValidator is a no-op validator which only emits warning string when fieldValue
-// is set to the specified one
-func noopValueWarningValidator(fieldValue interface{}, warningText string) schema.SchemaValidateFunc {
-	return func(i interface{}, k string) (warnings []string, errors []error) {
-		if fieldValue == i {
-			warnings = append(warnings, fmt.Sprintf("%s\n\n", warningText))
-		}
-
-		return
-	}
-}
-
-// anyValueWarningValidator is a validator which only emits always warning string
-func anyValueWarningValidator(fieldValue interface{}, warningText string) schema.SchemaValidateFunc {
-	return func(i interface{}, k string) (warnings []string, errors []error) {
-		warnings = append(warnings, fmt.Sprintf("%s\n\n", warningText))
-		return
-	}
-}
-
-func checkEmptyOrSingleIP() schema.SchemaValidateFunc {
-	return func(i interface{}, k string) (s []string, es []error) {
-		v, ok := i.(string)
-		if !ok {
-			es = append(es, fmt.Errorf("expected type of %s to be string", k))
-			return
-		}
-
-		if net.ParseIP(v) == nil && v != "" {
-			es = append(es, fmt.Errorf(
-				"expected %s to be empty or contain a valid IP, got: %s", k, v))
-		}
-		return
 	}
 }
